% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreppingModelFitAndJump.R
\name{AssignPTScores}
\alias{AssignPTScores}
\title{Assign Pseudotime Scores  (deprecated)}
\usage{
AssignPTScores(cosine_data, seurat_obj, column_name)
}
\arguments{
\item{cosine_data}{the transformed PCA matrix/dataframe we want to add pseudotime column too.}

\item{seurat_obj}{The seurat object to pull the metadata column about pseudotime from}

\item{column_name}{column name you want to use for pseudotime}
}
\value{
The input matrix with an addition column added for the pseudotime (PT)
}
\description{
This takes the pseudotime column and assigns it to the pseudotime (PT_score) column.
}
\examples{
\dontrun{Z<-AssignPTScores(Z,seurat_obj,column_name="Combined_Ordering")}
}
