% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChainFunctions.R
\name{BatchCorrectGraph}
\alias{BatchCorrectGraph}
\title{Batch Correct Graph}
\usage{
BatchCorrectGraph(
  graph,
  meta,
  state_grouping_column_name,
  batch_correction_column_name
)
}
\arguments{
\item{graph}{The graph to be used for the correction, either will be visitation or CMFPT}

\item{meta}{the meta data columns of the seurat object to access the states and batch corrected columns}

\item{state_grouping_column_name}{the column name to be the states you determine your model by, often it is cell type metadata in seurat.}

\item{batch_correction_column_name}{the column that contains sample/mouse/patient differences.}
}
\value{
it returns the graph which should overfit the seurat object.
}
\description{
This function is used to batch correct the input graph either Visitation or CMFPT for each batch. Then it is expanded back out to the graph form.
}
\examples{
  \dontrun{CMFPT_sparse<-BatchCorrectGraph(CMFPT_sparse,meta,state_grouping_column_name,batch_correction_column_name)}
}
