% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChainFunctions.R
\name{BatchCorrectLineages}
\alias{BatchCorrectLineages}
\title{Batch Correct Lineages}
\usage{
BatchCorrectLineages(
  matrix_to_correct,
  metadata,
  state_grouping_column_name,
  batch_correction_column_name,
  verbose = T
)
}
\arguments{
\item{matrix_to_correct}{This can either be the Lineage fates or Visitation Matrix}

\item{metadata}{This is a dataframe with metadata for the state grouping and batches}

\item{state_grouping_column_name}{the clusters to summarize and aggregate by}

\item{batch_correction_column_name}{the column name to determine how the matrices are subsetted.}

\item{verbose}{updating where in batch correction process the fit is.}
}
\value{
This returns a lineages by cells matrix with batch corrected values renormalized between 0 and 1
}
\description{
This function performs mean matching batch correction on lineages in a similar fashion as ComBAT.
}
\examples{
 \dontrun{lineage_corrected<-BatchCorrectLineages(lineages_matrix,meta,state_grouping_column_name="custom_cell_classes_fine",batch_correction_column_name="sample")}
}
