% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoisedCellFunctions.R
\name{ClassifyPoisedCellsSingleLineage}
\alias{ClassifyPoisedCellsSingleLineage}
\title{Individual lineage and cluster for preferentially biasing}
\usage{
ClassifyPoisedCellsSingleLineage(data_subset, lineages_to_compare)
}
\arguments{
\item{data_subset}{This is the raw data subsetted on the initial cluster we want to investigate.}

\item{lineages_to_compare}{this is the data columns to look into, we use 3 of them: lineage fate, and then membership scores for current population and lineage population}
}
\value{
this returns a dataframe for each input cell with it's nominal outlier score, bias score, and hard cutoff value for biased.
}
\description{
This function finds the skewness for a cluster to handle which is an outlier vs inlier
}
\examples{
\dontrun{poised_vector<-ClassifyPoisedCellsSingleLineage(data_subset = data_input\%>\%
filter(.data[[state_grouping_column_name]]==as.character(unique(data_input[[state_grouping_column_name]]))[cluster_iter]),
c(lineages_to_compare[lineage_iter],
 paste0(current_group,"_membership"),
 paste0(lineages_to_compare[lineage_iter],"_membership")))
}
}
