% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreppingModelFitAndJump.R
\name{ClusterAssignColumn}
\alias{ClusterAssignColumn}
\title{Cluster Assignment for Cells (deprecated)}
\usage{
ClusterAssignColumn(cosine_data, seurat_obj, column_name)
}
\arguments{
\item{cosine_data}{The transformed PCA matrix we want to add another column to}

\item{seurat_obj}{the seurat object we want to pull information from (i.e. cell type or group)}

\item{column_name}{The column name we want to get out of the seurat object.}
}
\value{
The input matrix with an addition column added for the cell type group.
}
\description{
# get group assignments from the seurat object post seurat work
}
\examples{
\dontrun{Z<-ClusterAssignColumn(Z,seurat_obj,column_name="custom_cell_classes_fine")}
}
