% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChainFunctions.R
\name{ClusterToClusterFateMatrix}
\alias{ClusterToClusterFateMatrix}
\title{Turning cell by lineage matrix, to a cluster by lineage matrix}
\usage{
ClusterToClusterFateMatrix(seurat_obj, state_grouping_column_name)
}
\arguments{
\item{seurat_obj}{the seurat object with the Fates assay.}

\item{state_grouping_column_name}{the clusters to summarize and aggregate by}
}
\value{
It produces a cluster by cluster matrix by simply aggregating all the cell states in rows and columns
}
\description{
This is a utility function for future use.
}
\examples{
\dontrun{small_heatmap<-ClusterToClusterFateMatrix(seurat_obj,state_grouping_column_name="custom_cell_classes_fine")
pheatmap::pheatmap(small_heatmap,scale="row")
}
}
