% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreppingModelFitAndJump.R
\name{CombinePseudoOrdering}
\alias{CombinePseudoOrdering}
\title{Combining Pseudo Ordering with Entropy Efficiency
This strategy deploys a}
\usage{
CombinePseudoOrdering(seurat_obj, terminal_state_list)
}
\arguments{
\item{seurat_obj}{This is the seurat object with the scores for the lineages of interest}

\item{terminal_state_list}{this is the list of terminal lineages set earlier. Names are irrelevant. We just take the size of the terminal lineages to grab the last N score columns in the metadata}
}
\value{
This returns another column in the metadata of your seurat object called Combined_Ordering
}
\description{
Combining Pseudo Ordering with Entropy Efficiency
This strategy deploys a
}
\examples{
\dontrun{
Idents(seurat_obj)<-"custom_cell_classes_fine"
DefaultAssay(seurat_obj)<-"RNA"
markers.differences<-FindAllMarkers(seurat_obj)
terminal_state_list<-c("early_Erp","MkP","Lymph","Mast", "Neu", "cDC", "Baso")
reduced_markers<-markers.differences\%>\%filter(cluster\%in\%terminal_state_list)
gene_list_from_findallmarkers<-list()
for(iter in 1:length(terminal_state_list)){
  gene_list_from_findallmarkers[[iter]]<-reduced_markers\%>\%
    filter(p_val_adj<0.01)\%>\%
       filter(avg_log2FC>1)\%>\%
           filter(cluster==terminal_state_list[iter])\%>\%pull(gene)\%>\%unique
    }
seurat_obj<-GetPseudoOrdering(seurat_obj,gene_list_from_findallmarkers,terminal_state_list)
seurat_obj<-seurat_obj<-CombinePseudoOrdering(seurat_obj,terminal_state_list)
}

}
