% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreppingModelFitAndJump.R
\name{CosineTransformPCAData}
\alias{CosineTransformPCAData}
\title{Cosine Transform for PCA Data}
\usage{
CosineTransformPCAData(seurat_obj)
}
\arguments{
\item{seurat_obj}{Put in the Seurat object. Be sure PCA in seurat has been ran first since we pull the cell embeddings from the pca layer.}
}
\value{
It returns a matrix that transforms PCA into cosine space
}
\description{
cosine transforming data gives better representation than euclidean distance. We sorta cheat creating cosine dist with the norm below.
}
\examples{
\dontrun{seurat_obj<-CosineTransformPCAData(seurat_obj)
state_grouping_column_name="custom_cell_classes_fine"
pseudotime_column_name = "Combined_Ordering"
batch_correction_column_name="sample"
Z <-seurat_obj[["transformed_pca"]]@cell.embeddings\%>\%
as.data.frame\%>\%
tibble::rownames_to_column(var="row_names")\%>\%
dplyr::inner_join(Seurat::FetchData(seurat_obj,vars=c(state_grouping_column_name,pseudotime_column_name,batch_correction_column_name))\%>\%
tibble::rownames_to_column(var="row_names"),by="row_names")\%>\%
tibble::column_to_rownames(var="row_names")\%>\%
dplyr::ungroup()\%>\%
dplyr::arrange(.data[[pseudotime_column_name]])}
}
