% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPMFunctions.R
\name{extract_fit_array}
\alias{extract_fit_array}
\title{Extract the model fit parameters into a matrix given sinks and PCs}
\usage{
extract_fit_array(model_fits, nsink, npc)
}
\arguments{
\item{model_fits}{the large list of model filts.}

\item{nsink}{total number of sink cells we have}

\item{npc}{The total number of PCs we have (commonly 30-50)}
}
\value{
A matrix of all parameters from the fitted values.
}
\description{
Extract the model fit parameters into a matrix given sinks and PCs
}
\examples{
fit_array <- extract_fit_array(model_fits, nsink = nsink, npc = npc)

}
