% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChainFunctions.R
\name{FateMatrixByClusterMembership}
\alias{FateMatrixByClusterMembership}
\title{Fate Matrix scaled by cell clusters}
\usage{
FateMatrixByClusterMembership(seurat_obj)
}
\arguments{
\item{seurat_obj}{the seurat object with a TPM assay and metadata containing a column for absorbing states}
}
\value{
this returns a matrix that is cells by cluster.
}
\description{
This is a utility function for future development. It allows for a metric called committment
}
\examples{
 \dontrun{cell_to_clust_commitment<-FateMatrixByClusterMembership(seurat_obj)}
}
