% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoisedCellFunctions.R
\name{GetAllClassifiedPoisedCells}
\alias{GetAllClassifiedPoisedCells}
\title{Get Preferentially Biased Cells}
\usage{
GetAllClassifiedPoisedCells(
  seurat_obj,
  state_grouping_column_name,
  lineages_to_compare
)
}
\arguments{
\item{seurat_obj}{this must contain dataframe must contain membership and}

\item{lineages_to_compare}{this is the row names of the lineages_fates assay you wish to include in the study}
}
\value{
it returns a new assay called bias_scores. These are continuous values that dictate a score to use.
}
\description{
This function determines within a cluster group which cells are selectively biased toward the lineages selected. Important to note, we develop an assay we provide both an outlier score in data slot, and a bias score(recommended) in the scale.data slot.
}
\examples{
\dontrun{seurat_obj<-GetAllClassifiedPoisedCells(seurat_obj,state_grouping_column_name="custom_cell_class_fine",lineages_to_compare=c("Mono","Neu","Ery"))}
}
