% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbsorbingStateFunctions.R
\name{GetAutomaticAbsorbingCellAssignment}
\alias{GetAutomaticAbsorbingCellAssignment}
\title{Get a model based approach to determine Absorbing States}
\usage{
GetAutomaticAbsorbingCellAssignment(
  seurat_obj,
  markers_df = NULL,
  sink_cell_names = NULL,
  mode = c("markers", "CMFPT"),
  state_grouping_column_name,
  pseudotime_column_name,
  batch_correction_column_name,
  terminal_state_list
)
}
\arguments{
\item{seurat_obj}{the seurat object with a TPM assay and potentially the CMFPT}

\item{markers_df}{place the  is used if the mode="markers"}

\item{sink_cell_names}{a short cut to manually select cell names to be absorbing states default is NULL}

\item{mode}{if sink cell names is NULL, this selects which approach to use for absorbing states. currently choose 1 of 2 methods: "markers", "CMFPT"}

\item{state_grouping_column_name}{the column name to be the states you determine your model by, often it is cell type metadata in seurat.}

\item{pseudotime_column_name}{the column name to be the pseudotime, often it is Combined_Ordering.}

\item{batch_correction_column_name}{the column that contains sample/mouse/patient differences.}

\item{terminal_state_list}{this is a character vector specifying which cell states within state_grouping_column_name to keep (terminal states)}
}
\value{
seurat object with an updated metadata column called absorbing_states. NA means not selected, while 1 is selected.
}
\description{
This is a function that identifies the a relatively balanced absorbing state populations for each sample.
}
\examples{
\dontrun{
randomly_selected_cell_names<-sample(colnames(subset(seurat_obj,subset= cluster_annotations\%in\%terminal_state_list)),200,replace = F)
seurat_obj<-GetAutomaticAbsorbingCellAssignment(seurat_obj,sink_cell_names = randomly_selected_cell_names)
}

}
