% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChainFunctions.R
\name{GetBatchCorrectedCMFPT}
\alias{GetBatchCorrectedCMFPT}
\title{Wrapper for Batch Corrected Conditional Mean First Passage Time (CMFPT)}
\usage{
GetBatchCorrectedCMFPT(
  seurat_obj,
  state_grouping_column_name,
  batch_correction_column_name,
  old_flag = F,
  targets_in_rows = T,
  n_threads = 1,
  verbose = T
)
}
\arguments{
\item{seurat_obj}{The seurat object for each batch to determine the CMFPT}

\item{state_grouping_column_name}{the column name for the states, this is used for batch correction.}

\item{batch_correction_column_name}{the column name to determine how to subset cell-cell networks with TPMs}

\item{old_flag}{temporary flag for comparing new methods with one that is slow but mathematically exact.}

\item{targets_in_rows}{This should be left as TRUE, but optional control for futture downstream changes,}

\item{n_threads}{left as 1 for safety, consider setting higher based on computer hardware. 8 on an M2 Macbook air took 21hours on about 7500 cells. Further parallel strategies to improve speed will eventually be coming soon.}

\item{verbose}{Informs whether to update which sample/batch we are currently working on.}
}
\value{
should return a graphs object called CMFPT in the seurat object
}
\description{
This function is the wrapper to the log10 of Conditional Mean First Passage Time (CMFPT). It will perform CMFPT for each batch sequentially. This function will take a long time, so it is advised to
For reference, on an M2 Macbook air with 8 threads took 21 hours for 7500 cells. More bench marking will be conducted to improve the speed of this calculation.
The Conditional Mean First Passage Time (CMFPT) is a measure that answers how quickly cell i(row) will visit cell j(column) on average before ending up in an absrobing state.
The measure is the number of time steps. It is important to batch correct as this is a network size dependent measure, it is ill-advised to compare head to head.
}
\examples{
 \dontrun{seurat_obj<-GetBatchCorrectedCMFPT(seurat_obj,batch_correction_column_name="sample",old_flag = F,targets_in_rows = T,n_threads = 8,verbose=T)}
}
