% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChainFunctions.R
\name{GetCellLineageFateMatrix}
\alias{GetCellLineageFateMatrix}
\title{Aggregating the Fates from absorbing states to lineages}
\usage{
GetCellLineageFateMatrix(
  Bmat,
  seurat_obj,
  absorbing_state_column_name,
  state_grouping_column_name
)
}
\arguments{
\item{Bmat}{This is the fate matrix that was found using GetFateMatrix}

\item{seurat_obj}{Used so we can extract the two metadata columns below cleanly.}

\item{absorbing_state_column_name}{T This is a metadata column that contains TRUE if it is absorbing state, NA is set for transient cells, used to ensure we get correct names to aggregate}

\item{state_grouping_column_name}{This is the grouping function to aggregate the absorbing probabilities by}
}
\value{
It returns a dense matrix of n_cells by l_lineages
}
\description{
The goal of this function is to aggregate the absorbing states by their state columns.
}
\examples{
lineage_fate_matrix<-GetCellLineageFateMatrix(Bmat,seurat_obj,absorbing_state_column_name="absorbing_state_bool",state_grouping_column_name="custom_cell_classes_fine")
}
