% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPMFunctions.R
\name{GetCellToCellTPM_fast}
\alias{GetCellToCellTPM_fast}
\title{Wrapper Function for the cell-to-cell Transition Probability Matrix (TPM) with batch correction and parallelization}
\usage{
GetCellToCellTPM_fast(
  seurat_obj,
  Y,
  model_fits,
  state_grouping_column_name,
  pseudotime_column_name,
  batch_correction_column_name,
  n_cores = NULL,
  cell_indices_to_use = NULL,
  order = 5,
  eps = 9.99999999999999e-301,
  normalize_rows = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{seurat_obj}{the main seurat object to analyze.}

\item{Y}{The sink cells identified from GetSinkData}

\item{model_fits}{The model fit equations from Reduced_Jump_Prob_fast}

\item{state_grouping_column_name}{the column name to be the states you determine your model by, often it is cell type metadata in seurat.}

\item{pseudotime_column_name}{the column name to be the pseudotime, often it is Combined_Ordering.}

\item{batch_correction_column_name}{the column that contains sample/mouse/patient differences.}

\item{n_cores}{this is how many parallel threads to run at once. If using 8,}

\item{cell_indices_to_use}{set to NULL, index if only interested in a particular cell.}

\item{order}{this is a truncated value to consider the poisson process in the infinite sum.}

\item{eps}{stability value for logprob transform.}

\item{normalize_rows}{set this to true to turn into a markov process, default is true.}

\item{verbose}{if want status updates on samples.}
}
\value{
A seurat object with an updated graphs called TPM.
}
\description{
Wrapper Function for the cell-to-cell Transition Probability Matrix (TPM) with batch correction and parallelization
}
\examples{
 \dontrun{seurat_obj<-GetCellToCellTPM_fast(seurat_obj,Y,Prob_reduced,state_grouping_column_name="custom_cell_classes_fine",pseudotime_column_name = "Combined_Ordering",batch_correction_column_name="sample",n_cores = 8,cell_indices_to_use = NULL,order = 5,eps = 1e-300,normalize_rows = TRUE,verbose = TRUE)}

}
