% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChainFunctions.R
\name{GetConditionalMeanFirstPassageTime}
\alias{GetConditionalMeanFirstPassageTime}
\title{Get Conditional Mean First Passage Time in R parallel strategy}
\usage{
GetConditionalMeanFirstPassageTime(
  seurat_obj,
  old_flag = F,
  targets_in_rows = T,
  n_threads = 1
)
}
\arguments{
\item{seurat_obj}{The seurat object for each batch to determine the CMFPT}

\item{old_flag}{temporary flag for comparing new methods with one that is slow but mathematically exact.}

\item{targets_in_rows}{This should be left as TRUE, but optional control for futture downstream changes,}

\item{n_threads}{left as 1 for safety, consider setting higher based on computer hardware. 8 on an M2 Macbook air took 21hours on about 7500 cells. Further parallel strategies to improve speed will eventually be coming soon.}
}
\value{
This will return a single square matrix containing how long on average it takes for cell i(row) to visit cell j (column)
}
\description{
The Conditional Mean First Passage Time (CMFPT) is a measure that answers how quickly cell i(row) will visit cell j(column) on average before ending up in an absrobing state.
The measure is the number of time steps. It is important to batch correct as this is a network size dependent measure, it is ill-advised to compare head to head.
}
\examples{
 \dontrun{Sparse_CMFPT <- GetConditionalMeanFirstPassageTime(subset(seurat_obj, cells = batch_cells),old_flag=old_flag,targets_in_rows=targets_in_rows,n_threads=n_threads)}
}
