% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChainFunctions.R
\name{GetFateMatrix}
\alias{GetFateMatrix}
\title{Calculate Fate and Fundamental Matrices}
\usage{
GetFateMatrix(seurat_obj, absorbing_state_column_name)
}
\arguments{
\item{seurat_obj}{The seurat object, please check you have TPM as a graph name.}

\item{absorbing_state_column_name}{This is a metadata column that contains TRUE if it is absorbing state, NA is set for transient cells}
}
\value{
This will return a list of 2 matrices, a fate matrix B, and a fundamental matrix N.
}
\description{
This function is the main workhorse we use absorbing markov chains and partition the TPM  into a Q (transient matrix) and R (absorbing matrix)
Afterwards, we solve for the fundamental matrix which is N = (I-Q)^{-1} and find the Fate matrix through B=NR
}
\examples{
sparse_fate_matrices<-GetFateMatrix(seurat_obj,absorbing_state_column_name="absorbing_state_bool")

}
