% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChainFunctions.R
\name{GetFateMatrixAndMetrics}
\alias{GetFateMatrixAndMetrics}
\title{Getting the Fate Matrix as an Assay}
\usage{
GetFateMatrixAndMetrics(
  seurat_obj,
  state_grouping_column_name,
  pseudotime_column_name,
  batch_correction_column_name,
  absorbing_state_column_name,
  verbose = T
)
}
\arguments{
\item{seurat_obj}{the seurat object with a TPM assay and metadata containing a column for absorbing states}

\item{state_grouping_column_name}{the column name to be the states you determine your model by, often it is cell type metadata in seurat.}

\item{pseudotime_column_name}{the column name to be the pseudotime, often it is Combined_Ordering.}

\item{batch_correction_column_name}{the column that contains sample/mouse/patient differences.}

\item{absorbing_state_column_name}{T This is a metadata column that contains TRUE if it is absorbing state, NA is set for transient cells, used to ensure we get correct names to aggregate}

\item{verbose}{informs which batch we are currently working.}
}
\value{
It will return a seurat object with updated assays, and graphs.
}
\description{
The main wrapper function to place Fates into their own assay, with scaled counts as lineages. The lineage Fates will contain batch corrected values.
The Visitation probability will be batch corrected and designated as a graph in the seurat object
The CMFPT takes a long time to run, and most analysis does not require this, however, to run the CMFPT measure please run the GetConditionalMeanFirstPassageTime,
Once visitation and CMFPT graphs are identified one can run GetWeightedDestinationTime to obtain the WDT which is also stored as a graph.
}
\examples{
seurat_obj<-GetFateMatrixAndMetrics(seurat_obj,state_grouping_column_name="custom_cell_classes_fine",pseudotime_column_name="Combined_Ordering",batch_correction_column_name="sample",absorbing_state_column_name="absorbing_states")
}
