% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFWrapper.R
\name{GetFateTFModel}
\alias{GetFateTFModel}
\title{Get Basic multivariate linear model of fates and and TFs}
\usage{
GetFateTFModel(seurat_obj, lineages_to_use)
}
\arguments{
\item{seurat_obj}{The seurat object which contains lineage fates and the tfsulm assay}

\item{lineages_to_use}{this is a list of the rownames to use in the lineage_fates assay for the model fit.}
}
\value{
it returns a dataframe containing the statistical results of comparison between the lineage_fates and TFs assay
}
\description{
This model is used to correlate fates with TFs to nominate candidate hypotheses for mechanisms driving the fates.
}
\examples{
model_fit_df<-GetFateTFModel(seurat_obj,c("B_cells","Neu","Mono","Ery"))
}
