% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduced_Jump_Prob.R
\name{GetJumpProgramFeatures}
\alias{GetJumpProgramFeatures}
\title{Get a table and genes that associated with Jumps}
\usage{
GetJumpProgramFeatures(
  seurat_obj,
  Y,
  model_fits_list,
  state_grouping_column_name = "custom_cell_classes_fine",
  pseudotime_column_name = "Combined_Ordering",
  batch_correction_column_name = "sample"
)
}
\arguments{
\item{seurat_obj}{the seurat object to identify the jumps}

\item{Y}{A set of data points for sink points from GetEigenData}

\item{state_grouping_column_name}{the column name to be the states you determine your model by, often it is cell type metadata in seurat.}

\item{pseudotime_column_name}{the column name to be the pseudotime, often it is Combined_Ordering.}

\item{batch_correction_column_name}{the column that contains sample/mouse/patient differences.}

\item{model_fits}{The model fit equations from Reduced_Jump_Prob_fast}
}
\value{
it returns a list of two entries, 1) is the list of PCs and state clusters that have jumps. 2) is the list of candidate genes to look into that significant contributors to the PC.
}
\description{
Important to note, we identify genes from the principle components not the transformed PCs
}
\examples{
PC_genes_on_jump_processes<-GetJumpProgramFeatures(seurat_obj,Y,model_fits_list, state_grouping_column_name="custom_cell_classes_fine",pseudotime_column_name="Combined_Ordering",batch_correction_column_name="sample")

}
