% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChainFunctions.R
\name{GetMembership}
\alias{GetMembership}
\title{Get Membership of cells through gaussian estimation}
\usage{
GetMembership(seurat_obj, state_grouping_column_name)
}
\arguments{
\item{seurat_obj}{this is the input seurat object to incorporate. It needs the column group to function}

\item{state_grouping_column_name}{the column name to be the states you determine your model by, often it is cell type metadata in seurat.}
}
\value{
It returns a seurat object assay of membership assignments that are normalized per group
}
\description{
Get Membership of cells through gaussian estimation
}
\examples{
\dontrun{seurat_obj<-GetMembership(seurat_obj,state_grouping_column_name="custom_cell_classes_fine")}
}
