% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreppingModelFitAndJump.R
\name{GetPseudoOrdering}
\alias{GetPseudoOrdering}
\title{Get Pseudo Ordering for each geneset list.}
\usage{
GetPseudoOrdering(seurat_obj, supervised_list_of_list, name_of_score)
}
\arguments{
\item{seurat_obj}{is the seurat object you wish to get pseudo time ordering of}

\item{supervised_list_of_list}{This is either a list of gene names, or a list of lists of gene names you want scores for}

\item{name_of_score}{Individual scores can have names}
}
\value{
it returns the seurat object
}
\description{
Note that this function will handle matching the names for the genesets to the same format as the seurat rows. This will work regardless human or mouse.
}
\examples{
\dontrun{
Idents(seurat_obj)<-"custom_cell_classes_fine"
DefaultAssay(seurat_obj)<-"RNA"
markers.differences<-FindAllMarkers(seurat_obj)
terminal_state_list<-c("early_Erp","MkP","Lymph","Mast", "Neu", "cDC", "Baso")
reduced_markers<-markers.differences\%>\%filter(cluster\%in\%terminal_state_list)
gene_list_from_findallmarkers<-list()
for(iter in 1:length(terminal_state_list)){
  gene_list_from_findallmarkers[[iter]]<-reduced_markers\%>\%
    filter(p_val_adj<0.01)\%>\%
       filter(avg_log2FC>1)\%>\%
           filter(cluster==terminal_state_list[iter])\%>\%pull(gene)\%>\%unique
    }
seurat_obj<-GetPseudoOrdering(seurat_obj,gene_list_from_findallmarkers,terminal_state_list)
seurat_obj<-seurat_obj<-CombinePseudoOrdering(seurat_obj,terminal_state_list)
}

}
