% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreppingModelFitAndJump.R
\name{GetSinkData}
\alias{GetSinkData}
\title{Get Sink data points}
\usage{
GetSinkData(
  seurat_obj,
  state_grouping_column_name = "custom_cell_classes_fine",
  pseudotime_column_name = "Combined_Ordering"
)
}
\arguments{
\item{seurat_obj}{The seurat object with a transformed PC space.}

\item{state_grouping_column_name}{the column name to be the states you determine your model by, often it is cell type metadata in seurat.}
}
\value{
A set of data points for eigen cells used for the jump-drift-diffusion model
}
\description{
this is a summarized group for points and is used for model fits. It will pull all groups that listed as states.
}
\examples{
\dontrun{
Y<-GetSinkData(seurat_obj,state_grouping_column_name="custom_cell_classes_fine",pseudotime_column_name="Combined_Ordering")
}
}
