% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChainFunctions.R
\name{GetTransientStateMeasures}
\alias{GetTransientStateMeasures}
\title{Get Visitation Probability and Expected timesteps for TPM}
\usage{
GetTransientStateMeasures(Nmat)
}
\arguments{
\item{Nmat}{this is the fundamental matrix found before}
}
\value{
this returns a list of measures including timesteps to any absorbing state and visitation probability.
}
\description{
This function obtains metrics for transient to transient state networks. It contains 3 measures for now. The first is the expected time to
}
\examples{
 \dontrun{vistation_metrics<-GetTransientStateMeasures(sparse_fate_list[["Nmat"]])}

}
