% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChainFunctions.R
\name{GetWeightedDestinationTime}
\alias{GetWeightedDestinationTime}
\title{Weighted Destination Time (WDT) function}
\usage{
GetWeightedDestinationTime(
  seurat_obj,
  preference_lineage_column = NULL,
  enrichment_flag = T
)
}
\arguments{
\item{seurat_obj}{The seurat object for each batch to determine the CMFPT}

\item{preference_lineage_column}{Which meta data column to appropriately aggregate for a reduced representation, leave NULL if it not interested.}

\item{enrichment_flag}{Set TRUE if you want the enrichment of the preference lineage column to a null distribution.}
}
\value{
the seurat object with an updated Assay called WDT and 2 new matrices in seurat_obj misc layer
}
\description{
This function requires that GetMembership, GetBatchCorrectedCMFPT, and GetFateMatrixAndMetrics were run first.
This takes the
}
\examples{
 \dontrun{seurat_obj<-GetWeightedDestinationTime(seurat_obj,preference_lineage_column="sgRNA",enrichment_flag=T)}
}
