% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPMFunctions.R
\name{JDD_sum_over_sinks_blocked}
\alias{JDD_sum_over_sinks_blocked}
\title{Blocked Structured to solve Jump drift diffusion, helpful for massive amount of cells in Seurat object}
\usage{
JDD_sum_over_sinks_blocked(
  pars_mat,
  Xt,
  Yp,
  deltat,
  y_dts,
  X0,
  t0,
  order = 5,
  block_size = 100,
  eps_sd = 1e-12
)
}
\arguments{
\item{pars_mat}{fitted parameters matrix from the Reduced_Jump_Prob}

\item{Xt}{is all cells we are transitioning towards.}

\item{Yp}{is the sink/eigen cells we use for particular PC}

\item{deltat}{this is the absolute value for the pseudo time from the starting cell to all other cells}

\item{y_dts}{this is the pseudo time of the sink cells Yp}

\item{X0}{initial starting cell we are finding transition probabilities from.}

\item{t0}{initial starting pseudo time.}

\item{order}{this is a truncated value to consider the poisson process in the infinite sum.}

\item{eps_sd}{just a stabilizing control}
}
\value{
it returns a probability density of the transitions from a starting cell.
}
\description{
Blocked Structured to solve Jump drift diffusion, helpful for massive amount of cells in Seurat object
}
\examples{
\dontrun{
for (PC_iter in seq_len(npc)) {
pars_mat <- fit_array[, PC_iter, , drop = FALSE]
pars_mat <- matrix(pars_mat, nrow = nsink, ncol = 4)
prob_per_pc[, PC_iter] <- JDD_sum_over_sinks_blocked(pars_mat   = pars_mat,
Xt         = Xmat[target_idx, PC_iter],
Yp         = Ymat[, PC_iter],
 deltat     = deltat,
 y_dts      = ypt,
 X0         = Xmat[starting_point, PC_iter],
  t0         = t0,
  order      = order,
  block_size = 100)
  }}

}
