% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPMFunctions.R
\name{JDD_sum_over_sinks_vec}
\alias{JDD_sum_over_sinks_vec}
\title{This is a summation over the Jump drift diffusion model.}
\usage{
JDD_sum_over_sinks_vec(
  pars_mat,
  Xt,
  Yp,
  deltat,
  y_dts,
  X0,
  t0,
  order = 5,
  eps_sd = 1e-12
)
}
\arguments{
\item{pars_mat}{fitted parameters matrix from the Reduced_Jump_Prob}

\item{Xt}{is all cells we are transitioning towards.}

\item{Yp}{is the sink/eigen cells we use for particular PC}

\item{deltat}{this is the absolute value for the pseudo time from the starting cell to all other cells}

\item{y_dts}{this is the pseudo time of the sink cells Yp}

\item{X0}{initial starting cell we are finding transition probabilities from.}

\item{t0}{initial starting pseudo time.}

\item{order}{this is a truncated value to consider the poisson process in the infinite sum.}

\item{eps_sd}{just a stabilizing control}
}
\value{
probability density of each state Xt
}
\description{
This is a summation over the Jump drift diffusion model.
}
\examples{
\dontrun{
for (idx in blocks) {
out <- out + JDD_sum_over_sinks_vec(
  pars_mat = pars_mat[idx, , drop = FALSE],
  Xt       = Xt,
  Yp       = Yp[idx],
  deltat   = deltat,
  y_dts    = y_dts[idx],
  X0       = X0,
  t0       = t0,
  order    = order,
  eps_sd   = eps_sd
  )
  }}
}
