% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduced_Jump_Prob.R
\name{JumpROPE}
\alias{JumpROPE}
\title{Perform Gene Ontology to find the Relevant Ontology Program Extraction (JumpROPE)}
\usage{
JumpROPE(seurat_obj, jump_gene_list)
}
\arguments{
\item{seurat_obj}{the seurat object to identify the jumps}

\item{jump_gene_list}{the output list from GetJumpProgramFeatures that contains genes from PCs}
}
\value{
A list of 2 gene ontology sets for each PC where at least a single jump is located.
}
\description{
Important to note, we identify genes from the principle components not the transformed PCs
}
\examples{
GO_terms_in_jump_PCs<-JumpROPE(seurat_obj,jump_gene_list)

}
