% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reduced_Jump_Prob.R
\name{Reduced_Jump_Prob_fast}
\alias{Reduced_Jump_Prob_fast}
\title{Fitting Jump Model for each Eigen Cell and PC.}
\usage{
Reduced_Jump_Prob_fast(
  seurat_obj,
  Total_Eigen_Data,
  start_point,
  state_grouping_column_name = "custom_cell_classes_fine",
  pseudotime_column_name = "Combined_Ordering",
  batch_correction_column_name = "sample",
  order = 5,
  maxeval = 1000,
  xtol_rel = 1e-05
)
}
\arguments{
\item{seurat_obj}{the main seurat object to analyze.}

\item{Total_Eigen_Data}{A set of data points for sink points from GetEigenData}

\item{start_point}{index for which cell is "origin" or earliest cell}

\item{state_grouping_column_name}{the column name to be the states you determine your model by, often it is cell type metadata in seurat.}

\item{pseudotime_column_name}{the column name to be the pseudotime, often it is Combined_Ordering.}

\item{batch_correction_column_name}{the column that contains sample/mouse/patient differences.}

\item{order}{this is a truncated value to consider the poisson process in the infinite sum.}

\item{maxeval}{the total number of evaluations for the Negative log-likelihood fit (default is 1000)}

\item{xtol_rel}{the relative tolerance for the nloptr default is 1e-5}
}
\value{
It returns a list of models for each EigenCell and PC
}
\description{
Fitting Jump Model for each Eigen Cell and PC.
}
\examples{
\dontrun{
seurat_obj<-CosineTransformPCAData(seurat_obj)
Y <- GetSinkData(seurat_obj,state_grouping_column_name="custom_cell_classes_fine",pseudotime_column_name = "Combined_Ordering")
model_fits <- Reduced_Jump_Prob_fast(seurat_obj,Y,start_point=1,state_grouping_column_name="custom_cell_classes_fine",pseudotime_column_name="Combined_Ordering",batch_correction_column_name="sample")


}
}
