% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarkovChainFunctions.R
\name{removeIllConditionedCells}
\alias{removeIllConditionedCells}
\title{Check for removing non-connected cells, necessary for Fate and Visitation}
\usage{
removeIllConditionedCells(TPM_matrix)
}
\arguments{
\item{TPM_matrix}{The transition probability matrix.}
}
\value{
It returns a truncated TPM matrix with the problematic cells eliminated in both rows and columns.
}
\description{
We use this to ensure our graph eliminates singleton cells
}
\examples{
  \dontrun{TPM_matrix<-removeIllConditionedCells(TPM_matrix)}
}
