% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TFWrapper.R
\name{TFWrapper}
\alias{TFWrapper}
\title{Wrapper function to obtain Transcription Factor Activity}
\usage{
TFWrapper(seurat_obj, assay_name = "RNA", organism = "mouse", model = "TF")
}
\arguments{
\item{seurat_obj}{the main seurat object to analyze.}

\item{assay_name}{This should be kept at "RNA", however, newer models may allow "SCT"}

\item{organism}{this is either "mouse" or "human".}

\item{model}{this is chosen between "TF", "progeny", or "hallmark"}
}
\value{
the same seurat object with a new assay for TFs
}
\description{
Wrapper function to obtain Transcription Factor Activity
}
\examples{
seurat_obj<-TFWrapper(seurat_obj,assay_name="RNA",organism="mouse",model="TF")
}
